% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_table_reference_geography}
\alias{create_table_reference_geography}
\title{Title}
\usage{
create_table_reference_geography(
  conn = get_db_conn(),
  postcodes = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{conn}{a database connection}

\item{postcodes}{a character vector of postcodes. If `NULL` (the default),
data will be fetched from the `reference_homes` table.}

\item{overwrite}{whether to overwrite the `reference_geography` table.
Default is `TRUE`.}
}
\value{

}
\description{
Title
}
