% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_table_reference_homes}
\alias{create_table_reference_homes}
\title{Create or overwrite the \code{reference_homes} table}
\usage{
create_table_reference_homes(
  homes = reference_homes,
  conn = get_db_conn(),
  overwrite = TRUE
)
}
\arguments{
\item{conn}{a database connection}

\item{overwrite}{whether to overwrite the table (default is TRUE)}
}
\value{
the name of the table if data was successfully loaded
}
\description{
Create or overwrite the \code{reference_homes} table
}
