% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{funnel_outlier}
\alias{funnel_outlier}
\title{Classify observation as an outlier based on Poisson control limits}
\usage{
funnel_outlier(o, denom, fun, confidence_limit = 0.9)
}
\arguments{
\item{o}{the numeric of the observed count}

\item{denom}{a numeric value for the rate denominator (exposed to risk/person-days)}

\item{fun}{a function generated by \code{\link{funnel_poisson_control_limits}()}}

\item{confidence_limit}{a numeric value between 0 and 1 for the confidence level
for outlier detection (eg 0.95)}
}
\value{

}
\description{
Classify observation as an outlier based on Poisson control limits
}
