% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{km_ct_estimator}
\alias{km_ct_estimator}
\title{Kaplan-Meier survivor function point and interval estimator for count-time data}
\usage{
km_ct_estimator(df, t, d, pop, subtract_cases = TRUE, overwrite = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{t}{unquoted name of a time variable}

\item{d}{unquoted name of the count of events}

\item{pop}{unquoted name of the population (inclusive of those who previously
experienced the event, but removing those that are censored, e.g. those who
left the study without experiencing the event)}

\item{subtract_cases}{a boolean indicated whether events count `d` should be
substracted from pop as time passes in order to calculate the total exposed-to-risk.
The default is `TRUE` for when `pop` number includes person-days for past cases.}

\item{overwrite}{a boolean indicating whether the function can overwrite variables
`n_t`, `S_t` and `S_t_SE` in `df` (default is TRUE)}
}
\value{
the input data frame `df` with additional variables `n_t` for the population
at risk size, `S_t` for the survivor function point estimator, and `S_t_SE` for its
standard error. A further `S_t_SElog` corresponding to an asymptotic estimator of the
standard deviation, which can be used to caculate confident interval bounds like so:
\itemize{
  \item \code{St_min = S_t^exp(-1.96*S_t_SElog)}
  \item \code{St_max = S_t^exp(+1.96*S_t_SElog)}
}
}
\description{
Kaplan-Meier estimator based on Greenwods' formula
}
\references{
section 1.1.4 of: Kalbfleisch JD, Prentice RL. Statistical Analysis of
Failure Time Data. New York: Wiley 1980.
}
