% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_data}
\alias{load_data}
\title{Load data extracts into the global environment}
\usage{
load_data()
}
\value{
nothing
}
\description{
Loads data extracts stored in directory `getOption("fshc_files")`:
`"residents.rda"`, `"incident.rda"`, `"occupancy.rda"`, `"beds.rda"`,
`"reference_homes.rda"`, `"reference_geography.rda"`
}
