% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{min_non_missing}
\alias{min_non_missing}
\title{Minimum non missing value}
\usage{
min_non_missing(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector
}
\description{
Unlike \code{\link{min}()}, this function returns NA
if there is no non-missing element in `x`
}
