% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{plot_funnel}
\alias{plot_funnel}
\title{Create funnel plot with Poisson}
\usage{
plot_funnel(data, y, denom, home_code, home_name)
}
\arguments{
\item{data}{a data frame}

\item{y}{the unquoted name of a numeric variable containing observed counts in `data`}

\item{denom}{the unquoted name of a numeric variable containing the denominator in `data`}

\item{home_code}{the unquoted name of a character variable containing the care home code in `data`}

\item{home_name}{the unquoted name of a character variable containing the care home name in `data`}
}
\value{
An object of class ggplot
}
\description{
Create funnel plot with Poisson
}
