% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{resident_days_approx_indicator}
\alias{resident_days_approx_indicator}
\title{Approximate presence in care home}
\usage{
resident_days_approx_indicator(data)
}
\arguments{
\item{data}{a data frame containing varibles `date`, `original_admission_date`,
`last_admission_date`, `last_discharge_date`, `date_end`, and `status`}
}
\value{
a data.frame with an rday variable indicating whether the resident
was in or out of the care me on the date contained in variable `date`
}
\description{
Approximate presence in care home
}
