% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_table_reference_geography}
\alias{create_table_reference_geography}
\title{Create or overwrite \code{reference_geography} table}
\usage{
create_table_reference_geography(
  conn = get_db_conn(),
  data = reference_geography,
  overwrite = TRUE
)
}
\arguments{
\item{conn}{a database connection}

\item{data}{a data frame, set to  `reference_geography` by default.}

\item{overwrite}{whether to overwrite the `reference_geography` table.
Default is `TRUE`.}
}
\value{

}
\description{
Create or overwrite \code{reference_geography} table
}
